added styles in assessment to seperate file
