+++ Aura_frontend/app/index.tsx
@@ -7,7 +7,6 @@ const Index = () => {
   useEffect(() => {
     const checkAuth = async () => {
       const token = await AsyncStorage.getItem("token");
-
       if (token) {
         router.replace("/(tabs)/Home");
       } else {
